+++ src/main/java/com/nhnacademy/book/member/domain/repository/auth/MemberAuthRepository.java
@@ -1,13 +1,10 @@
 package com.nhnacademy.book.member.domain.repository.auth;
 
-import com.nhnacademy.book.member.domain.Auth;
 import com.nhnacademy.book.member.domain.MemberAuth;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 
 @Repository
