+++ src/main/java/com/nhnacademy/book/member/domain/service/MemberAuthService.java
@@ -1,58 +1,14 @@
 package com.nhnacademy.book.member.domain.service;
 
-import com.nhnacademy.book.member.domain.Auth;
-import com.nhnacademy.book.member.domain.Member;
-import com.nhnacademy.book.member.domain.MemberAuth;
+
 import com.nhnacademy.book.member.domain.dto.auth.MemberAuthAssignRequestDto;
 import com.nhnacademy.book.member.domain.dto.auth.MemberAuthAssignResponseDto;
-import com.nhnacademy.book.member.domain.repository.auth.AuthRepository;
-import com.nhnacademy.book.member.domain.repository.auth.MemberAuthRepository;
-import com.nhnacademy.book.member.domain.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-@Service
-@RequiredArgsConstructor
-public class MemberAuthService {
-    private MemberAuthRepository memberAuthRepository;
-    private MemberRepository memberRepository;
-    private AuthRepository authRepository;
 
-    @Transactional
-    public String assignAuthToMember(MemberAuthAssignRequestDto requestDto) {
-        Long memberId = requestDto.getMemberId();
-        Long authId = requestDto.getAuthId();
-
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new IllegalArgumentException("회원이 존재하지 않습니다."));
-        Auth auth = authRepository.findById(authId)
-                .orElseThrow(() -> new IllegalArgumentException("권한이 존재하지 않습니다."));
-
-        // 새로운 멤버 권한 설정
-        MemberAuth memberAuth = new MemberAuth();
-        memberAuth.setMember(member);
-        memberAuth.setAuth(auth);
-
-        memberAuthRepository.save(memberAuth);
-        return "회원 권한이 성공적으로 설정되었습니다.";
-    }
-
-    public List<MemberAuthAssignResponseDto> getMemberAuthsByMemberId(Long memberId) {
-        return memberAuthRepository.findByMemberMemberId(memberId).stream()
-                .map(auth -> new MemberAuthAssignResponseDto(
-                        auth.getMemberAuthId(),
-                        auth.getMember().getMemberId(),
-                        auth.getAuth().getAuthId()))
-                .collect(Collectors.toList());
-    }
-
-    @Transactional
-    public void deleteMemberAuthsByMemberId(Long memberId) {
-        memberAuthRepository.deleteByMemberMemberId(memberId);
-    }
+public interface MemberAuthService {
+    String assignAuthToMember(MemberAuthAssignRequestDto requestDto);
+    List<MemberAuthAssignResponseDto> getMemberAuthsByMemberId(Long memberId);
+    void deleteMemberAuthsByMemberId(Long memberId);
 }
-
